+++ common/com_model.h
@@ -12,7 +12,7 @@
 #define STUDIO_RENDER 1
 #define STUDIO_EVENTS 2
 
-#define MAX_EDICTS 900
+#define MAX_EDICTS 2048
 
 #define MAX_MODEL_NAME 64
 #define MAX_MAP_HULLS 4